somedjangoproject/requirements.txt
Django==1.11.29
pytz==2022.7.1
logging==0.4.9.6
celery==4.4.7
faker==3.0.1
pillow==6.2.2
uwsgi==2.0.21
python-dotenv==0.18.0